# specs/001-rag-document-intelligence/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Prompt - Multi-Tenant AI-Powered RAG-Based Document Intelligence Platform
  description: API for document generation, RAG, authentication, and subscription management
  version: 1.0.0
servers:
  - url: https://api.prompt.com/v1
    description: Production server
  - url: https://staging-api.prompt.com/v1
    description: Staging server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRegistration:
      type: object
      required:
        - email
        - password
        - phone
        - domain_id
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          format: password
          minLength: 8
          description: User's password
        phone:
          type: string
          description: User's phone number in international format
        domain_id:
          type: string
          format: uuid
          description: ID of the domain the user is joining
    UserLogin:
      type: object
      required:
        - email_or_phone
        - password
      properties:
        email_or_phone:
          type: string
          description: User's email or phone number
        password:
          type: string
          format: password
          description: User's password
    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: JWT refresh token
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        phone:
          type: string
        domain_id:
          type: string
          format: uuid
        subscription_tier:
          type: string
          enum: [basic, pro, premium, institutional]
        role:
          type: string
          enum: [user, domain_admin, root_admin]
        created_at:
          type: string
          format: date-time
    Domain:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        icon_url:
          type: string
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
    ConversationRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: User query in English, Urdu, or Sindhi (auto-detected)
        language_hint:
          type: string
          enum: [english, urdu, sindhi]
          description: Optional language hint to override auto-detection
        response_language:
          type: string
          enum: [english, urdu, sindhi]
          description: Desired response language (defaults to same as query language)
    ConversationResponse:
      type: object
      properties:
        response:
          type: string
          description: Agent response in the requested language/script
        detected_language:
          type: string
          enum: [english, urdu, sindhi]
        response_language:
          type: string
          enum: [english, urdu, sindhi]
        rag_sources:
          type: array
          items:
            type: object
            properties:
              document_id:
                type: string
                format: uuid
              chunk_text:
                type: string
              confidence:
                type: number
        is_rag_grounded:
          type: boolean
          description: True if response is based on retrieved knowledge base content
    VoiceTranscribeRequest:
      type: object
      required:
        - audio
      properties:
        audio:
          type: string
          format: binary
          description: Audio file (WAV or M4A, max 25MB)
        language_hint:
          type: string
          enum: [english, urdu, sindhi]
          description: Optional language hint for better accuracy
    VoiceTranscribeResponse:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        transcription:
          type: string
          description: Transcribed text in original script (RTL for Urdu/Sindhi)
        detected_language:
          type: string
          enum: [english, urdu, sindhi]
        confidence:
          type: number
          minimum: 0
          maximum: 1
        low_confidence_warning:
          type: boolean
          description: True if confidence < 0.8 — user should review transcription
    SpeechSynthesisRequest:
      type: object
      required:
        - text
        - language
      properties:
        text:
          type: string
          description: Text to convert to speech (max 5000 chars)
        language:
          type: string
          enum: [english, urdu, sindhi]
    DocumentUpload:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
          description: PDF file to upload (max 10MB)
        document_type:
          type: string
          enum: [act, case_law, sample, textbook, protocol, standard]
          description: Type of document being uploaded
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        file_size_bytes:
          type: integer
        mime_type:
          type: string
        domain_id:
          type: string
          format: uuid
        document_type:
          type: string
          enum: [act, case_law, sample, textbook, protocol, standard]
        status:
          type: string
          enum: [pending, approved, rejected]
        uploaded_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    ChatRequest:
      type: object
      required:
        - message
        - template_id
      properties:
        message:
          type: string
          description: User's input for document generation
        template_id:
          type: string
          format: uuid
          description: ID of the template to use
        parameters:
          type: object
          additionalProperties: true
          description: Additional parameters for the template
    ChatResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
          description: Generated document content
        template_used:
          type: string
          format: uuid
        retrieved_sources:
          type: array
          items:
            type: object
            properties:
              document_id:
                type: string
                format: uuid
              chunk_id:
                type: string
                format: uuid
              confidence_score:
                type: number
                minimum: 0
                maximum: 1
              content:
                type: string
        validation_status:
          type: string
          enum: [valid, invalid, pending]
        output_formats:
          type: object
          properties:
            in_app:
              type: string
              description: Content formatted for in-app display
            pdf_url:
              type: string
              format: uri
              description: URL to download PDF version
            docx_url:
              type: string
              format: uri
              description: URL to download DOCX version
    Template:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        domain_id:
          type: string
          format: uuid
        description:
          type: string
        content:
          type: string
          description: Template content with {{slots}}
        slot_definitions:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
              required:
                type: boolean
              data_source:
                type: string
        formatting_rules:
          type: object
          additionalProperties: true
        version:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        tier:
          type: string
          enum: [basic, pro, premium, institutional]
        status:
          type: string
          enum: [active, expired, cancelled]
        start_date:
          type: string
          format: date
        expiry_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
    TokenApplication:
      type: object
      required:
        - token_code
      properties:
        token_code:
          type: string
          description: Promotional token code to apply
    TokenApplicationResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        benefit:
          type: object
          description: Details of the benefit applied

paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully, verification code sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: string
                    format: uuid
        '400':
          description: Invalid input data
        '409':
          description: Email or phone already exists

  /auth/verify:
    post:
      summary: Verify user registration with code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - verification_code
              properties:
                user_id:
                  type: string
                  format: uuid
                verification_code:
                  type: string
      responses:
        '200':
          description: Verification successful, login tokens returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid verification code
        '404':
          description: User not found

  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials

  /auth/refresh:
    post:
      summary: Refresh access token
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        '401':
          description: Invalid refresh token

  /domains:
    get:
      summary: Get all available domains
      responses:
        '200':
          description: List of available domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Domain'

  /documents/upload:
    post:
      summary: Upload a document to knowledge base
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUpload'
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Invalid file or exceeds size limit
        '401':
          description: Unauthorized
        '403':
          description: Insufficient permissions

  /documents:
    get:
      summary: Get user's uploaded documents
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
          description: Filter by document status
      responses:
        '200':
          description: List of user's documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'

  /chat:
    post:
      summary: Generate document using RAG and templates
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Document generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request or missing required data
        '401':
          description: Unauthorized
        '403':
          description: Insufficient permissions or subscription tier
        '422':
          description: Validation error - document could not be generated due to missing data

  /templates:
    get:
      summary: Get available templates for user's domain
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of available templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'

  /subscriptions/current:
    get:
      summary: Get user's current subscription
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current subscription details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'

  /conversation:
    post:
      summary: Send a multilingual conversation query to the domain agent
      description: >
        Accepts text input in English, Urdu, or Sindhi. Auto-detects language and
        returns a response in the same language/script. This is NOT free-form AI generation
        — responses are RAG-grounded and template-constrained.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationRequest'
      responses:
        '200':
          description: Agent response in same language as query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '400':
          description: Query validation failed
        '401':
          description: Unauthorized
        '422':
          description: Language detection failed or unsupported script

  /voice/transcribe:
    post:
      summary: Transcribe voice audio to text (Phase 2 – Pilot)
      description: >
        Accepts audio file (WAV or M4A, max 25MB). Returns transcription in the
        detected or specified language (English/Urdu/Sindhi) in correct script.
        Audio is deleted from server within 30 seconds of transcription completion.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VoiceTranscribeRequest'
      responses:
        '200':
          description: Transcription result with confidence score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceTranscribeResponse'
        '400':
          description: Invalid audio file or unsupported format
        '401':
          description: Unauthorized
        '422':
          description: Transcription failed — audio quality too low

  /voice/synthesize:
    post:
      summary: Convert text response to speech audio (Phase 3 – Full Scale)
      description: >
        Accepts text and language. Returns audio stream of the text spoken in
        English, Urdu, or Sindhi using language-native voice models.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeechSynthesisRequest'
      responses:
        '200':
          description: Audio stream (MP3)
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '503':
          description: TTS service unavailable

  /tokens/apply:
    post:
      summary: Apply a promotional token
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenApplication'
      responses:
        '200':
          description: Token applied successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenApplicationResponse'
        '400':
          description: Invalid or expired token
        '401':
          description: Unauthorized
        '409':
          description: Token already used by this user